Add skeleton for a python CLI script
I would be surprised if this was sufficient to install the CLI, but it
should be enough to run tests and start building out the interface
that you want.

To run the tests, `cd` to the "cron-between" directory,
`pip install -r requirements.txt`, and run `pytest`. Only one test,
based on hardcoded text, so should pass assuming croniter installs.

The name "cron-between" and that the CLI would be called "between"
is just me guessing at something. Call it what you will. :)